fe/fe-core/src/test/java/com/starrocks/epack/failover/ReplicatedObjectMgrTest.java
// Copyright 2021-present StarRocks, Inc. All rights reserved.

package com.starrocks.epack.failover;

import com.starrocks.epack.sql.ast.CreatePrimaryFailoverGroupStmt;
import com.starrocks.server.CatalogMgr;
import com.starrocks.server.GlobalStateMgr;
import com.starrocks.sql.analyzer.AnalyzeTestUtil;
import com.starrocks.sql.ast.CreateTableStmt;
import com.starrocks.utframe.StarRocksAssert;
import com.starrocks.utframe.UtFrameUtils;
import org.junit.Assert;
import org.junit.BeforeClass;
import org.junit.Test;

import static com.starrocks.sql.analyzer.AnalyzeTestUtil.analyzeSuccess;

public class ReplicatedObjectMgrTest {
    private static StarRocksAssert starRocksAssert;

    @BeforeClass
    public static void beforeClass() throws Exception {
        UtFrameUtils.createMinStarRocksCluster();
        AnalyzeTestUtil.init();
        starRocksAssert = new StarRocksAssert(AnalyzeTestUtil.getConnectContext());
        starRocksAssert.withDatabase("test").useDatabase("test");

        String sql = "create table single_partition_duplicate_key (key1 int, key2 varchar(10))\n" +
                "distributed by hash(key1) buckets 1\n" +
                "properties('replication_num' = '1'); ";
        CreateTableStmt createTableStmt = (CreateTableStmt) UtFrameUtils.parseStmtWithNewParser(sql,
                AnalyzeTestUtil.getConnectContext());
        GlobalStateMgr.getCurrentState().getLocalMetastore().createTable(createTableStmt);
    }

    @Test
    public void testDiff() throws Exception {
        CreatePrimaryFailoverGroupStmt stmt = (CreatePrimaryFailoverGroupStmt) analyzeSuccess(
                "CREATE FAILOVER GROUP test_group1 " +
                        "CATALOGS = default_catalog " +
                        "MEMBERS = " +
                        "'az1:SELF'," +
                        "'az2:192.168.0.1:9090'" +
                        "SCHEDULE = '1h'");

        ReplicatedObjectMgr objectMgr = new ReplicatedObjectMgr(stmt);
        ReplicatedObjectMeta objectMeta = objectMgr.saveToObjectMeta();

        for (ReplicatedObjectMeta.CatalogMeta catalogMeta : objectMeta.getCatalogMetas().values()) {
            Assert.assertTrue(CatalogMgr.isInternalCatalog(catalogMeta.getCatalogId()));
            Assert.assertTrue(CatalogMgr.isInternalCatalog(catalogMeta.getCatalogName()));
            Assert.assertTrue(!catalogMeta.getDatabases().isEmpty());
        }
    }
}
